feat: move project into modules, add coverage and package publish
