+++ api-client/src/main/scala/com/matthewjones372/api/client/HttpClientConfig.scala
@@ -1,0 +1,8 @@
+package com.matthewjones372.api.client
+
+import zio.http.URL
+
+final case class HttpClientConfig(
+  baseUrl: URL,
+  cacheSize: Int
+)
