+++ build.sbt
@@ -1,27 +1,101 @@
-ThisBuild / version := "0.1.0-SNAPSHOT"
+import sbtdynver.DynVerPlugin.autoImport._
 
-ThisBuild / scalaVersion := "3.4.2"
+ThisBuild / organization         := "com.matthewjones372"
+ThisBuild / name                 := "starwars-api"
+ThisBuild / organizationHomepage := Some(url("https://github.com/matthewjones372"))
+ThisBuild / scalaVersion         := "3.4.2"
+//version                          := dynver.value
+
+ThisBuild / publishTo := {
+  val gh = s"https://maven.pkg.github.com/matthewjones372/starwars-api"
+  if (isSnapshot.value)
+    Some("GitHub Package Registry" at gh)
+  else
+    Some("GitHub Package Registry" at gh)
+}
+
+ThisBuild / credentials += Credentials(
+  "GitHub Package Registry",
+  "maven.pkg.github.com",
+  sys.env.getOrElse("GITHUB_ACTOR", ""),
+  sys.env.getOrElse("GITHUB_TOKEN", "")
+)
+
+publishMavenStyle := true
+
+dynverVTagPrefix                    := false // No v-prefix in the version tags
+ThisBuild / dynverSonatypeSnapshots := true
+
+ThisBuild / testFrameworks    := Seq(TestFramework("zio.test.sbt.ZTestFramework"))
+ThisBuild / publish / skip    := true
+ThisBuild / publishMavenStyle := true
+Global / onChangedBuildSource := ReloadOnSourceChanges
+
+lazy val oneToOneClassMapping = "test->test;compile->compile"
 
 lazy val root = (project in file("."))
   .settings(
-    name             := "swapi",
-    idePackagePrefix := Some("com.jones")
+    name := "swapi"
+  )
+  .dependsOn(
+    modules.map(_ % oneToOneClassMapping): _*
+  )
+  .aggregate(modules: _*)
+
+lazy val domain = Projects
+  .create("domain")
+  .settings(
+    Libraries.zioSchema,
+    Libraries.zioJson,
+    Libraries.zioTest
   )
 
-libraryDependencies ++= Seq(
-  "dev.zio"              %% "zio"               % "2.1.6",
-  "dev.zio"              %% "zio-http"          % "3.0.0-RC9",
-  "dev.zio"              %% "zio-json"          % "0.7.1",
-  "dev.zio"              %% "zio-prelude"       % "1.0.0-RC27",
-  "dev.zio"              %% "zio-schema-json"   % "1.3.0",
-  "dev.zio"              %% "zio-cache"         % "0.2.3",
-  "dev.zio"              %% "zio-concurrent"    % "2.1.6",
-  "nl.vroste"            %% "rezilience"        % "0.9.4",
-  "dev.zio"              %% "zio-http-testkit"  % "3.0.0-RC9" % Test,
-  "dev.zio"              %% "zio-test"          % "2.1.5"     % Test,
-  "dev.zio"              %% "zio-test-sbt"      % "2.1.6"     % Test,
-  "dev.zio"              %% "zio-test-magnolia" % "2.1.6"     % Test,
-  "io.github.kitlangton" %% "stubby"            % "0.1.2"     % Test
-)
+lazy val client = Projects
+  .create("api-client")
+  .settings(
+    Libraries.zioHttp,
+    Libraries.zioCache,
+    Libraries.resilience,
+    Libraries.zioTest
+  )
+  .settings(
+    publish / skip := false
+  )
+  .dependsOn(
+    domain % oneToOneClassMapping
+  )
 
-testFrameworks += new TestFramework("zio.test.sbt.ZTestFramework")
+lazy val data = Projects
+  .create("data")
+  .settings(
+    Libraries.zio,
+    Libraries.zioJson,
+    Libraries.zioTest
+  )
+  .dependsOn(
+    domain % oneToOneClassMapping
+  )
+
+lazy val `http-api` = Projects
+  .create("http-api")
+  .settings(
+    Libraries.zio,
+    Libraries.zioHttp,
+    Libraries.zioTest
+  )
+  .dependsOn(
+    domain % oneToOneClassMapping,
+    data   % oneToOneClassMapping
+  )
+
+lazy val search = Projects
+  .create("search")
+  .settings(
+    Libraries.zio,
+    Libraries.zioTest
+  )
+  .dependsOn(
+    domain % oneToOneClassMapping
+  )
+
+lazy val modules: Seq[ProjectReference] = Seq(domain, client, data, `http-api`, search)
