+++ data/src/main/scala/com/matthewjones372/data/DataRepoError.scala
@@ -1,0 +1,8 @@
+package com.matthewjones372.data
+
+enum DataRepoError(msg: String) extends RuntimeException(msg):
+  case PersonNotFound(message: String, personId: Int) extends DataRepoError(s"Person with id $personId not found.")
+  case FilmNotFound(message: String, filmId: Int) extends DataRepoError(s"Film with id $filmId not found.")
+  case UnexpectedError(message: String, exception: Exception)
+      extends DataRepoError(s"Data-Repo has encountered an unexpected error: $message. Exception: $exception")
+  case FilmsNotFound extends DataRepoError(s"No films found.")
