+++ domain/src/main/scala/com/matthewjones372/domain/Paged.scala
@@ -1,0 +1,13 @@
+package com.matthewjones372.domain
+
+trait Paged[A]:
+  def count: Int
+  def results: List[A]
+
+  final def pageCount: Int =
+    if results.isEmpty then 0
+    else {
+      val fullPages      = count / results.length
+      val hasPartialPage = if (count % results.length == 0) 0 else 1
+      fullPages + hasPartialPage
+    }
