+++ project/plugins.sbt
@@ -1,3 +1,6 @@
-addSbtPlugin("org.jetbrains.scala"       % "sbt-ide-settings"   % "1.1.2")
-addSbtPlugin("org.scalameta"             % "sbt-scalafmt"       % "2.4.6")
-addSbtPlugin("io.github.davidgregory084" % "sbt-tpolecat"       % "0.4.1")
+addSbtPlugin("org.jetbrains.scala"       % "sbt-ide-settings"    % "1.1.2")
+addSbtPlugin("org.scalameta"             % "sbt-scalafmt"        % "2.4.6")
+addSbtPlugin("io.github.davidgregory084" % "sbt-tpolecat"        % "0.4.1")
+addSbtPlugin("com.github.sbt"           %% "sbt-native-packager" % "1.9.16")
+addSbtPlugin("org.scoverage"            %% "sbt-scoverage"       % "2.0.9")
+addSbtPlugin("com.github.sbt"            % "sbt-dynver"          % "5.0.1")
