--- src/main/scala/client/HttpClientConfig.scala
@@ -1,9 +1,0 @@
-package com.jones
-package client
-
-import zio.http.URL
-
-final case class HttpClientConfig(
-  baseUrl: URL,
-  cacheSize: Int
-)
